fix: created utils to remove code duplication
