+++ packages/crayons-styles/src/index.js
@@ -1,30 +1,5 @@
 import crayons from '../dist/design-tokens/crayons.json';
-const base = '$base.';
-
-function getValue(value) {
-  if (value.startsWith(base)) {
-    const baseValue = value
-      .replace(base, '')
-      .split('.')
-      .reduce((p, c) => {
-        return p + '-' + c;
-      }, '--fw');
-    return `var(${baseValue})`;
-  }
-  return value;
-}
-
-function convertTokenToString(tokens) {
-  let cssString = '';
-  for (const [, value] of Object.entries(tokens)) {
-    if ('var' in value) {
-      cssString = cssString + `${value['var']} : ${getValue(value['value'])};`;
-    } else {
-      cssString = cssString + convertTokenToString(value);
-    }
-  }
-  return cssString;
-}
+import { convertTokenToString } from './utils';
 
 function setStyle(tokens) {
   const cssString = convertTokenToString(tokens);
